test: 반복 횟수 검증 테스트 재작성
- 검증 로직이 Controller로 이전하였으므로 그에 맞춰 테스트코드를 이동한다.
- 반복 횟수가 유효하지 않을때 다시 입력받는지 확인하는 테스트 코드를 추가한다.
